Added basic 1995 Crash support and octree rendering

+++ CrashEdit/Controls/SceneryEntryViewer.cs
@@ -95,7 +95,7 @@ namespace CrashEdit
             {
                 GL.Color3(Color.Fuchsia);
             }
-            GL.Vertex3((double)(entry.XOffset + vertex.X * 16),(double)(entry.YOffset + vertex.Y * 16),(double)(entry.ZOffset + vertex.Z * 16));
+            GL.Vertex3(entry.XOffset + vertex.X * 16,entry.YOffset + vertex.Y * 16,entry.ZOffset + vertex.Z * 16);
         }
 
         protected override void Dispose(bool disposing)
